+++ src/app/data/db.service.ts
@@ -1,5 +1,5 @@
 import { Injectable, signal } from '@angular/core';
-import { Event, Day, Camp, Art, Pin, DataMethods, MapSet, GeoRef, Dataset, RSLEvent, TimeRange } from './models';
+import { Event, Day, Camp, Art, Pin, DataMethods, MapSet, GeoRef, Dataset, RSLEvent, TimeRange, GPSSet } from './models';
 import { call, registerWorker } from './worker-interface';
 import { daysUntil, noDate, now } from '../utils/utils';
 import { GpsCoord, Point } from '../map/geo.utils';
@@ -83,6 +83,10 @@ export class DbService {
 
   public async getMapPoints(name: string): Promise<MapSet> {
     return await call(this.worker, DataMethods.GetMapPoints, name);
+  }
+
+  public async getGPSPoints(name: string, title: string): Promise<MapSet> {
+    return await call(this.worker, DataMethods.GetGPSPoints, name, title);
   }
 
   public async getRSL(terms: string, day: Date | undefined, gpsCoord: GpsCoord | undefined): Promise<RSLEvent[]> {